Robustify the error handling
I ran libmo through the afl fuzzer, and it showed that the error cases
were pretty woeful; there were multiple segfaults.

Now we check that the file doesn't contain insane contents - and abort
if so. Most codepaths report errors back to their callers now, with the
notable exception of MoGroup - it's unclear to me how or whether to
handle that.